requirements.txt

Django==4.2.7
psycopg2-binary==2.9.7
gunicorn==21.2.0
djangorestframework==3.14.0
